#authorization

+++ laracasts/30-days-to-learn-laravel/app/Http/Controllers/JobController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
-use Illuminate\Http\Request;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -15,14 +17,17 @@ class JobController extends Controller
             'jobs' => $jobs
         ]);
     }
+
     public function create()
     {
         return view('jobs.create');
     }
+
     public function show(Job $job)
     {
         return view('jobs.show', ['job' => $job]);
     }
+
     public function store()
     {
         request()->validate([
@@ -39,12 +44,16 @@ class JobController extends Controller
         return redirect('/jobs');
 
     }
+
     public function edit(Job $job)
     {
         return view('jobs.edit', ['job' => $job]);
     }
+
     public function update(Job $job)
     {
+        Gate::authorize('edit', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required']
@@ -58,8 +67,11 @@ class JobController extends Controller
         return redirect('/jobs/' . $job->id);
 
     }
+
     public function destroy(Job $job)
     {
+        Gate::authorize('edit', $job);
+
         $job->delete();
         return redirect('/jobs');
     }
